requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.16.0
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
yarl==1.19.0
